Rename HttpUpgrade.[cc,h] files to HttpUpgradeProtocolAccess.[cc,h]
To follow squid coding guidelines. The HttpUpgradeProtocolAccess is
the primary class in these files the other class are in practice
sub-classes or helper classes.
